examples/react_bash/pyproject.toml
[project]
name = "react-bash-example"
version = "0.1.0"
description = "ReAct Loop with Bash Tool — DeepSeek Reasoner"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "supervisors",
]

[tool.uv.sources]
supervisors = { path = "../../", editable = true }
